third_party/blink/web_tests/fast/canvas-api/canvas-overloads-strokeText.html

<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<body>
<script>
var ctx = document.createElement('canvas').getContext('2d');

test(function(t) {
    assert_throws_js(TypeError, function() {
      ctx.strokeText();
    });
    assert_throws_js(TypeError, function() {
      ctx.strokeText('moo');
    });
    assert_throws_js(TypeError, function() {
      ctx.strokeText('moo', 0);
    });
    assert_equals(ctx.strokeText('moo',0,0), undefined);
    assert_equals(ctx.strokeText('moo',0,0,0), undefined);
    assert_equals(ctx.strokeText('moo',0,0,0,0), undefined);
}, "Test the behavior of CanvasRenderingContext2D.strokeText() when called with different numbers of arguments.");
</script>
</body>
